pyproject.toml
[project]
name='evbets'
version='0.0.1'
dependencies=[
  'typer>=0.12.3',
  'httpx>=0.27.0',
  'orjson>=3.10.5',
  'polars>=1.5.0',
  'python-dotenv>=1.0.1',
  'scikit-learn>=1.5.0',
  'scipy>=1.13.0',
  'joblib>=1.4.2',
  'betacal>=1.1.0',
]

[project.scripts]
evbets='evbets.cli:app'

[build-system]
requires=['setuptools','wheel']
build-backend='setuptools.build_meta'
